+++ src/main/java/at/hannibal2/skyhanni/features/minion/MinionFeatures.kt
@@ -199,7 +199,7 @@ class MinionFeatures {
                 val lastEmptied = minion.value
                 if (lastEmptied == 0L) continue
                 val duration = System.currentTimeMillis() - lastEmptied
-                val format = StringUtils.formatDuration(duration / 1000)
+                val format = StringUtils.formatDuration(duration / 1000) + " ago"
                 if (LocationUtils.canSee(playerEyeLocation, location)) {
                     val text = "§eHopper Emptied: $format"
                     event.drawString(location.add(0.0, 2.0, 0.0), text, true)